+++ src/index.js
@@ -4,7 +4,7 @@ import './helpers/intercom';
 import { injectGlobal } from 'styled-components';
 import { globalStyles } from './styles';
 import { bootIntercom } from './helpers/utilities';
-import { getLocal } from './handlers/localstorage';
+import { getLanguage } from './handlers/localstorage';
 import lang, { resources } from './languages';
 import Root from './Root';
 
@@ -16,7 +16,8 @@ bootIntercom();
 
 // Languages (i18n)
 lang.init({
-  lng: getLocal('language') || 'en',
+  lng: getLanguage() || 'en',
+  fallbackLng: 'en',
   debug: process.env.NODE_ENV === 'development',
   resources,
 });